requirements.txt
prometheus-api-client==0.5.5
pandas>=2.0.0
google-genai>=0.3.0
pyyaml>=6.0
